LayoutTests/editing/selection/move-selection-backward-at-beginning-of-line-by-sentence-granularity.html

<!DOCTYPE html>
<html>
<body>
<div id="editor" contenteditable>
This is<br>
<div id="target">one sentence.</div>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests moving the selection backwards at the beginning of a line by sentence granularity. The caret should move to the first line.');

editor.focus();
getSelection().setPosition(target, 0);
getSelection().modify('move', 'backward', 'sentence');

Markup.dump(editor);

</script>
</body>
</html>
