LayoutTests/editing/selection/move-selection-backward-at-beginning-of-line-by-sentence-granularity-expected.txt

This tests moving the selection backwards at the beginning of a line by sentence granularity. The caret should move to the first line.
| "
<#selection-caret>This is"
| <br>
| "
"
| <div>
|   id="target"
|   "one sentence."
| "
"
